+++ actix-web-lab/src/html.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use actix_web::{
     http::{
         header::{self, ContentType, TryIntoHeaderValue},
@@ -7,6 +9,7 @@ use actix_web::{
 };
 
 /// An HTML responder.
+#[deprecated(since = "0.21.0", note = "Graduated to Actix Web.")]
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Html(pub String);